Create mixin for node properties instead of weird declaration stuff
